[modify][gws][circular] don't update `#updated` on set-seen, unset-seen, soft-delete and undo-delete (#4599)
* [modify] don't touch updated during setting seen

* [modify] show last updated

* [modify] don't touch updated with soft delete and undo delete

* [modify] refactor w/ RuboCop

* [fixup] don't touch updated with soft delete and undo delete

* [fixup] gws circular datetimes

+++ app/controllers/concerns/gws/crud_filter.rb
@@ -44,11 +44,9 @@ module Gws::CrudFilter
 
   def edit
     raise "403" unless @item.allowed?(:edit, @cur_user, site: @cur_site)
-    if @item.is_a?(Cms::Addon::EditLock)
-      unless @item.acquire_lock
-        redirect_to action: :lock
-        return
-      end
+    if @item.is_a?(Cms::Addon::EditLock) && !@item.acquire_lock
+      redirect_to action: :lock
+      return
     end
     render
   end
@@ -79,6 +77,7 @@ module Gws::CrudFilter
       return
     end
 
+    @item.record_timestamps = false
     @item.deleted = Time.zone.now
     render_destroy @item.save
   end
@@ -92,6 +91,7 @@ module Gws::CrudFilter
       return
     end
 
+    @item.record_timestamps = false
     @item.deleted = nil
 
     render_opts = {}
@@ -128,11 +128,9 @@ module Gws::CrudFilter
     entries.each do |item|
       if item.allowed?(:delete, @cur_user, site: @cur_site)
         item.attributes = fix_params
-        if item.is_a?(Gws::User)
-          if item.deletion_unlocked? && item.disabled?
-            item.destroy
-            next
-          end
+        if item.is_a?(Gws::User) && item.deletion_unlocked? && item.disabled?
+          item.destroy
+          next
         end
         next if item.disable
       else
@@ -155,6 +153,7 @@ module Gws::CrudFilter
       item.try(:cur_site=, @cur_site)
       item.try(:cur_user=, @cur_user)
       if item.allowed?(:delete, @cur_user, site: @cur_site)
+        item.record_timestamps = false
         item.deleted = Time.zone.now
         next if item.save
       else
